Fixed issue with incorrect border on drag and drop
Correct border is now show when dragging and dropping while the hex
config is open and after the border was changed.

Issue was caused by Token.clone (or rather PlaceableObject.clone) not
also duplicating the temporary flags correctly. This might be a better
way to fix the height and width issues as well actually