Add orientation flipping and additional sizing options.
This rewrites a couple bits of foundry's snapping logic to allow for
consistent orientation of size 2 and 4 tokens as well as enables support
for tokens larger than 4.

As a start, size 5 is added, which is a size that would snap wrong using
the default logic.

Global and token level orientation shifts are supported using a new
setting for global and the existing flag for token level.